Implement back button for view reports screen
