--- src/main/java/ReportsHomeScreenController.java
@@ -1,41 +1,0 @@
-import java.net.URL;
-import java.util.ResourceBundle;
-import javafx.fxml.FXML;
-import javafx.fxml.Initializable;
-import javafx.scene.control.Button;
-
-/**
- * FXML Controller class for reports home screen.
- */
-public class ReportsHomeScreenController implements Initializable {
-
-    @FXML
-    private Button backButton;
-    
-    @FXML
-    private Button bookingPercentagesReportButton;
-    
-    @FXML
-    private Button incomeReportButton;
-    
-    /**
-     * Initializes the controller class.
-     */
-    @Override
-    public void initialize(URL url, ResourceBundle rb) {
-        // TODO
-    }    
-    
-    public void onBackButtonClicke() {
-        // TODO
-    }
-    
-    public void onBookingPercentagsButtonClicked() {
-        // TODO
-    }
-    
-    public void onIncomeReportButttonClicked() {
-        // TODO
-    }
-    
-}
