+++ src/main/java/com/gitrekt/resort/controller/ReportsHomeScreenController.java
@@ -1,0 +1,54 @@
+package com.gitrekt.resort.controller;
+
+import java.io.IOException;
+import java.net.URL;
+import java.util.ResourceBundle;
+import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
+import javafx.fxml.Initializable;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.scene.control.Button;
+import javafx.stage.Stage;
+
+/**
+ * FXML Controller class for reports home screen.
+ */
+public class ReportsHomeScreenController implements Initializable {
+
+    @FXML
+    private Button backButton;
+    
+    @FXML
+    private Button bookingPercentagesReportButton;
+    
+    @FXML
+    private Button incomeReportButton;
+    
+    /**
+     * Initializes the controller class.
+     */
+    @Override
+    public void initialize(URL url, ResourceBundle rb) {
+        // TODO
+    }    
+    
+    public void onBackButtonClicked() throws IOException {
+        Stage mainStage = (Stage) backButton.getScene().getWindow();
+        Parent staffHomeScreenRoot = FXMLLoader.load(
+            getClass().getResource("/fxml/StaffHomeScreen.fxml")
+        );
+        Scene staffHomeScreen = new Scene(staffHomeScreenRoot);
+        mainStage.centerOnScreen();
+        mainStage.setScene(staffHomeScreen);
+    }
+    
+    public void onIncomeReportButtonClicked() {
+        // TODO
+    }
+    
+    public void onBookingPercentagesReportButtonClicked() {
+        // TODO
+    }
+    
+}
