+++ src/main/resources/fxml/ReportsHomeScreen.fxml
@@ -5,15 +5,14 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.VBox?>
 
-
-<AnchorPane stylesheets="@../fxcss/Master.css" xmlns="http://javafx.com/javafx/8.0.141" xmlns:fx="http://javafx.com/fxml/1">
+<AnchorPane stylesheets="@../fxcss/Master.css" xmlns="http://javafx.com/javafx/8.0.141" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.gitrekt.resort.controller.ReportsHomeScreenController">
    <children>
       <VBox alignment="CENTER" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="800.0" spacing="20.0" stylesheets="@../fxcss/Master.css" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
          <padding>
             <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
          </padding>
          <children>
-            <Button fx:id="bookingPercentagesReportButton" mnemonicParsing="false" onAction="#onBookingPercentagesReportButton" prefHeight="35.0" prefWidth="200.0" text="Booking Percentages Report">
+            <Button fx:id="bookingPercentagesReportButton" mnemonicParsing="false" onAction="#onBookingPercentagesReportButtonClicked" prefHeight="35.0" prefWidth="200.0" text="Booking Percentages Report">
                <VBox.margin>
                   <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
                </VBox.margin>
